refactor: move small infra modules to lib/infra/ (#7720)
Move agent, blob, compose, s3, session, session-history modules
into lib/infra/ directory as part of the lib/ restructuring epic.
Pure structural move with import path updates, no behavior changes.
