+++ turbo/apps/web/app/api/storages/commit/route.ts
@@ -17,7 +17,7 @@ import { resolveOrg } from "../../../../src/lib/org/resolve-org";
 import {
   s3ObjectExists,
   verifyS3FilesExist,
-} from "../../../../src/lib/s3/s3-client";
+} from "../../../../src/lib/infra/s3/s3-client";
 import { computeContentHashFromHashes } from "../../../../src/lib/infra/storage/content-hash";
 import { env } from "../../../../src/env";
 import { logger } from "../../../../src/lib/logger";
