+++ turbo/apps/web/app/api/webhooks/agent/storages/commit/route.ts
@@ -17,7 +17,7 @@ import { getOrgData } from "../../../../../../src/lib/org/org-cache-service";
 import {
   s3ObjectExists,
   verifyS3FilesExist,
-} from "../../../../../../src/lib/s3/s3-client";
+} from "../../../../../../src/lib/infra/s3/s3-client";
 import { computeContentHashFromHashes } from "../../../../../../src/lib/infra/storage/content-hash";
 import { env } from "../../../../../../src/env";
 import { logger } from "../../../../../../src/lib/logger";
