+++ turbo/apps/web/src/lib/infra/storage/upload-storage.ts
@@ -2,8 +2,8 @@ import { gzipSync } from "node:zlib";
 import { eq, and } from "drizzle-orm";
 import { VOLUME_ORG_USER_ID } from "@vm0/core";
 import { storages, storageVersions } from "../../../db/schema/storage";
-import { putS3Object, verifyS3FilesExist } from "../../s3/s3-client";
-import type { S3StorageManifest } from "../../s3/types";
+import { putS3Object, verifyS3FilesExist } from "../s3/s3-client";
+import type { S3StorageManifest } from "../s3/types";
 import { computeContentHashFromHashes, hashFileContent } from "./content-hash";
 import { createSingleFileTar } from "../../tar";
 import { env } from "../../../env";
