+++ turbo/apps/web/src/lib/slack/context.ts
@@ -1,6 +1,6 @@
 import type { WebClient } from "@slack/web-api";
 import { logger } from "../logger";
-import { uploadS3Buffer, generatePresignedUrl } from "../s3/s3-client";
+import { uploadS3Buffer, generatePresignedUrl } from "../infra/s3/s3-client";
 import { env } from "../../env";
 import { type SlackUserInfo, formatSenderBlock } from "./client";
