+++ turbo/apps/web/src/lib/zero/skills/sync-skills.ts
@@ -33,7 +33,7 @@ import {
   putS3Object,
   listS3Objects,
   deleteS3Objects,
-} from "../../s3/s3-client";
+} from "../../infra/s3/s3-client";
 import { skills } from "../../../db/schema/skill";
 import { storages, storageVersions } from "../../../db/schema/storage";
 import { env } from "../../../env";
