+++ turbo/apps/web/src/lib/zero/zero-compose-service.ts
@@ -2,7 +2,7 @@ import { eq, desc } from "drizzle-orm";
 import { agentComposes } from "../../db/schema/agent-compose";
 import { zeroAgents } from "../../db/schema/zero-agent";
 import { notFound } from "../errors";
-import { canAccessCompose } from "../agent/compose-access";
+import { canAccessCompose } from "../infra/agent/compose-access";
 import type { ComposeListItem } from "@vm0/core";
 
 /**
